# Improve normalizeURL performance