This also fixes the input object being modified by the function.

We avoid double iteration when processing an image list, and an unnecessary copy of the sitemap item object.

Also, ensure we return early when only a URL is given. This significantly improves performance in the `stream-2` perf test:

Before:

```
runs: 1 batches: 1 total: 1
testing lots of data
========= stream-2 =============
median: 13994.0±0.0ms
99th percentile: 13994.0ms
```

After:

```
runs: 1 batches: 1 total: 1
testing lots of data
========= stream-2 =============
median: 5575.0±0.0ms
99th percentile: 5575.0ms
```

Other benchmarks also see improvement:

Before:

```
runs: 10 batches: 10 total: 100
testing stream
========= stream =============
median: 1218.0±63.0ms
99th percentile: 1448.0ms
```

After:

```
runs: 10 batches: 10 total: 100
testing stream
========= stream =============
median: 1071.5±66.6ms
99th percentile: 1387.0ms
```

Use the "Hide whitespace" option in GitHub for clarity while reviewing.