@huntharo 2024-05-22
@mohd-akram - Thanks for the contribution!

Can you edit the summary with *why* the changes improve performance such as "eliminates date objects are being created and destroyed without being changed" or something to that effect?

Also, can you quantify at all the performance benefit on hundreds of thousands of items being written to files or to memory streams?  Is it 0.1% time reduction or 10% time reduction?  Any impact on memory usage or GC times?