@mohd-akram 2024-05-22
@huntharo Thanks for looking at this. I've added some more information in the description. There's no change in memory usage or GC times that I could find, there might a small improvement due reducing a copy, but not really the focus of the PR.