@derduher 2024-05-23
I'll note that I don't see the perf improvement on node 22 but do see it on the previous versions. Not sure what that's about.